+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 
 ## v2.5.3 (Unreleased)
 
+### 🆕 Features
+* Append `X-Backend-Server` header to proxy responses  
+
 ### 🛠 Bug fixes
 
 * Fix chain registration issue with Kaleido/Infura when multitenancy is enabled